Signals page v2.1

+++ grails-app/services/us/wearecurio/services/AnalyticsService.groovy
@@ -1,20 +1,29 @@
 package us.wearecurio.services
 
 import groovy.time.*
-import us.wearecurio.utility.Utils
 import us.wearecurio.model.AnalyticsTimeSeries
 import us.wearecurio.model.AnalyticsTagMembership
+import us.wearecurio.model.AnalyticsTask
 import us.wearecurio.model.TagProperties
 import us.wearecurio.model.Entry
 import us.wearecurio.model.User
 import us.wearecurio.model.Tag
+import us.wearecurio.server.BackgroundTask
+import us.wearecurio.utility.Utils
 
 import org.apache.commons.logging.LogFactory
 
 import grails.util.Environment
-import us.wearecurio.analytics.Interop
 
 class AnalyticsService {
+
+	private static DEBUG = false
+
+	public static def SERVERS = [
+		'127.0.0.1:8090',
+		'127.0.0.1:8091'
+	]
+
 	private static def log = LogFactory.getLog(this)
 	static transactional = false
 	public static def LOG_FILE_NAME = "/tmp/analytics.${(new Date()).format("YYYY-MM-DD_HH:mm")}.log".toString()
@@ -60,7 +69,6 @@ class AnalyticsService {
 	public static saveTagMembership(userId) {
 		def user = User.get(userId.toLong())
 		log.debug("saveTagMembership: " + userId + ", numgroups: " + user.getTagGroups().size)
-		//log.debug "Write Tag Group:: uid: " + userId + " num Groups: " + user.getTagGroups().size
 		user.getTagGroups().each { tagGroup ->
 			def tagGroupId = tagGroup.id
 			tagGroup.getTags(userId).each { tag ->
@@ -80,9 +88,19 @@ class AnalyticsService {
 		property.classifyAsEvent().save(flush:true)
 	}
 
-	public static processUser(userId) {
+	public static prepareUser(analyticsTask) {
+		def userId = analyticsTask.userId
+		if (null == userId) {
+			return null
+		}
 		def user = User.get(userId.toLong())
 		def tagIds = user.tags().collect { it.id }
+
+		if (DEBUG) {
+			if (tagIds.size > 10) {
+				tagIds = tagIds[0..10]
+			}
+		}
 
 		// Write out tagGroup -> tag join table.
 		log "user id ${userId}: SaveTagMembership(${userId})"
@@ -93,29 +111,32 @@ class AnalyticsService {
 			classifyProperty(TagProperties.createOrLookup(userId, tagId))
 		}
 
-		// Delete the whole caching table to avoid duplicates and orphaned
-		//	series of tags that have been completely deleted.
+		// Delete the user's realizations of fetchPlotData.
 		log "user id ${userId}: delete table analytics_time_series"
-		AnalyticsTimeSeries.executeUpdate('delete from AnalyticsTimeSeries')
+		AnalyticsTimeSeries.executeUpdate('delete from AnalyticsTimeSeries at where at.userId=?', [userId])
 
 		log "user id ${userId}: refreshSeriesCache(${userId})"
 		tagIds.each { tagId ->
 			refreshSeriesCache(userId, tagId)
 		}
+		log "user id ${userId}: refreshSeriesCache(${userId}) complete."
 
-		// Run analytics on user.
-		String environment = Environment.getCurrent().toString()
-		log "user id ${userId}: Interop.updateUser(\"${environment}\", ${userId}, ${LOG_FILE_NAME})"
-		Interop.updateUser(environment, userId, LOG_FILE_NAME)
+	}
+
+	public static processOneOfManyUsers(childTask) {
+		log.debug "AnalyticsTask processOneOfManyUsers @ ${childTask.serverAddress}: start"
+		prepareUser(childTask)
+		childTask.startProcessing()
 	}
 
 	public static processUsers() {
-		def user_ids = User.executeQuery('select u.id from User u order by u.id')
-		user_ids.each { userId	->
-			log.debug "user id ${userId}: AnalyticsService.processUser(${userId})"
-			processUser(userId)
-			log.debug "user id ${userId}:  finished AnalyticsService.processUser(${userId})"
-			log.debug "\n"
+		def parentTask = AnalyticsTask.createParent()
+		Utils.save(parentTask)
+		SERVERS.each { serverAddress ->
+			BackgroundTask.launch {
+				def childTask = AnalyticsTask.createChild(serverAddress, parentTask)
+				processOneOfManyUsers(childTask)
+			}
 		}
 	}
